RAG-KDD-Cup/FinanceRAG Challenge 2024/code/FinanceRAG/requirements.txt
datasets
python-dotenv
pydantic
pytrec_eval
typing_extensions
langchain_openai
numpy
nltk
openai
einops
triton
sentence-transformers
ninja
packaging
kaggle
torch==2.2.0
transformers==4.42.4
#flash-attn
